+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.41.0] (18/07/2025)
+Now it is possible to call update commands using the `--id` option, which allows to update a specific template by its ID.
+For example: `silverfin update-reconciliation --id 12345`
+
 ## [1.40.0] (08/07/2025)
 - `stats` command now displays the amount of yaml files that have at least two unit tests defined